refactor: 예외 처리 및 증복 로직 제거하기
