+++ src/main.hash.js
@@ -1,5 +1,5 @@
-import { mount } from "./core/render";
 import { createRouter } from "./core/router";
+import { mount } from "./core/render/mount.js";
 import { routes, onBeforeEach } from "./router.js";
 import { provide } from "./core/context";
